# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"poc-typescript-nestjs","description":"proof of concept","version":"v1","contact":{}},"tags":[{"name":"api","description":""}],"servers":[],"components":{"schemas":{"ResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":true},"message":{"type":"string","description":"Message of result","example":"Executado com sucesso!"}},"required":["result","message"]},"DataResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":true},"message":{"type":"string","description":"Message of result","example":"Executado com sucesso!"},"data":{"type":"object","description":"Data of result"}},"required":["result","message","data"]},"TaskResponseDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of task"}},"required":["description"]},"ExceptionResponseApiDto":{"type":"object","properties":{"result":{"type":"boolean","description":"Result of response","example":false},"message":{"type":"string","description":"Message of result","example":"BadRequest"}},"required":["result","message"]},"CreateTaskRequestDto":{"type":"object","properties":{"description":{"type":"string","description":"Descriptioin of task","example":""}},"required":["description"]},"UpdateTaskRequestDto":{"type":"object","properties":{"description":{"type":"string","description":"Descriptioin of task","example":""}},"required":["description"]}}},"paths":{"/":{"get":{"operationId":"AppController_getRunning","summary":"Simple get endpoint","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseApiDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseApiDto"}}}}}}},"/api/tasks":{"get":{"operationId":"TasksController_getAll","summary":"Get all tasks","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponseDto"}}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tasks"]},"post":{"operationId":"TasksController_createBadge","summary":"Create task","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TaskResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tasks"]}},"/api/tasks/{taskId}":{"get":{"operationId":"TasksController_getById","summary":"Get task by id","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TaskResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tasks"]},"put":{"operationId":"TasksController_updateBadge","summary":"Update task","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"result":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TaskResponseDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExceptionResponseApiDto"}}}}},"tags":["Tasks"]}}}}